==> 02_pre-processing.ipynb <==
# read csv

import pandas as pd

# input files
filename = "./sample data/feature.csv"
filename1 = "./sample data/wq0118_records.csv"
filename3 = "./sample data/wp0071_records.csv"
filename4 = "./sample data/R047_levels.csv"

filename2 = "./sample data/R047_R258_titration_records.csv"

# output directory
file_directory = './sample data/output_3_files.csv'
output_all_data = './sample data/output_all_data.csv'
outpuut_directory_all = './sample data/output_all_4_data.csv'
outpuut_directory_open = './sample data/output_open_4_data.csv'

# read csv
df_feature = pd.read_csv(filename, parse_dates=['timestamp'])
df1_0118 = pd.read_csv(filename1, parse_dates=['measurement_result_date'])
df3_0071 = pd.read_csv(filename3,parse_dates=[['Measurement Result Date','Minute 5']])
df4_047 = pd.read_csv(filename4, parse_dates=['measurement_result_date'], dayfirst=True)

# print shape
print ("feature (df) shape: ", df_feature.shape)
print ("wq0118 (df1) shape: ", df1_0118.shape)
print ("wp0071 (df3) shape: ", df3_0071.shape)
print ("R047 (df4) shape: ", df4_047.shape)

# is unique
print ("feature (df): ", df_feature['timestamp'].is_unique)
print ("wq0118 (df1): ", df1_0118['measurement_result_date'].is_unique)
print ("wp0071 (df3): ", df3_0071['Measurement Result Date_Minute 5'].is_unique)
print ("R047 (df4): ", df4_047['measurement_result_date'].is_unique)

### filter duplicated data 

#### wp0071 - pump

# tem_0071 = df3_0071[['Measurement Result Date_Minute 5','Average Result Value', 'Asset Code']].sort_values(['Measurement Result Date_Minute 5'])

# data = tem_0071[930000:930050]
# data.groupby(['Measurement Result Date_Minute 5']).agg({'Average Result Value': 'sum'})

# sum average result value
feature_pmp = df3_0071.groupby(['Measurement Result Date_Minute 5']).agg({'Average Result Value': 'sum'})
# feature_pmp['Average Result Value']

# change index to specific column
feature_pmp['Measurement_Result_Date_Minute_15'] = feature_pmp.index
feature_pmp.shape

#### New feature: Average Result Value - indicator for flow speed when pump open 

feature_pmp['Average Result Value'].describe()

feature_pmp['Measurement_Result_Date_Minute_15'].is_unique

#### features

features_32138 = df_feature.loc[df_feature['node_id']==31238]
features_32138['timestamp'].describe()

features_32138['timestamp'].is_unique

features_41908 = df_feature.loc[df_feature['node_id']==41908]
features_41908['timestamp'].describe()

### New feature: water level
- 25/09/2017

df4_047 = pd.read_csv('./sample data/R047_levels.csv', parse_dates=['measurement_result_date'],dayfirst=True)
df4_047['measurement_result_date'].describe()
# drop useless column
df4_047 = df4_047.drop(['statistic_type', 'asset_code', 'facility_code'], axis=1)
df4_047.columns = ['measurement_result_date', 'water_level']
df4_047 = df4_047.sort_values(['measurement_result_date'])

df4_047['measurement_result_date'].describe()

# merge (inner join)

# merge label and pump state
df_merge_1 = pd.merge(df1_0118, feature_pmp, left_on='measurement_result_date', right_on='Measurement_Result_Date_Minute_15',how='inner')
df_merge_1.shape

df_merge_1['measurement_result_date'].is_unique

# merge df with features where node_id = 32138 (same with 41908)
df_merge_3 = pd.merge(df_merge_1, features_32138, left_on='measurement_result_date', right_on='timestamp',how='inner')
df_merge_3.shape

df_merge_3['Average Result Value'].describe()

df_pump_open = df_merge_3.loc[df_merge_3['Average Result Value']>0]
df_pump_open.shape

df_pump_open.sort_values(['measurement_result_date']).head(5)

## output data by merging 4 files
- 25/09/2018

# merge 3 files with water level
df_merge_4 = pd.merge(df_merge_3, df4_047, left_on='measurement_result_date', right_on='measurement_result_date',how='inner')
print "Output of merging all 4 files:",df_merge_4.shape

df_merge_4['measurement_result_date'].sort_values().describe()

## output open data by filter all data of 4 files
- 25/09/2018

df_pump_open_4 = df_merge_3.loc[df_merge_3['Average Result Value']>0]
print "Output of opening data with water level:", df_pump_open_4.shape

# save file - 1

#### sort values by daytime and output


# 8641 - dataset in 3 months
df_merge_3.sort_values(['measurement_result_date']).to_csv(output_all_data,index=None)

# 2769 - dataset in 3 months only when pump is open
df_pump_open.sort_values(['measurement_result_date']).to_csv(file_directory, index=None)

## save file - 2
- 25/09/2018

# 8640 - dataset in 3 months from 4 files
df_merge_4.sort_values(['measurement_result_date']).to_csv(outpuut_directory_all, index=None)

# 2769 - dataset in 3 months only when pump is open
df_pump_open_4.sort_values(['measurement_result_date']).to_csv(outpuut_directory_open, index=None)

# plot 

x_axis= df_pump_open['measurement_result_date']
y_label = df_pump_open['average_result_value']
y_zone_cl = df_pump_open['zone_flow_hist_0min']
y_res_cl = df_pump_open['res_cl_hist_0min']

%matplotlib
import matplotlib.pyplot as plt

fig = plt.figure(figsize=(20, 10))

ax1 = fig.add_subplot(111)
ax1.plot(x_axis, y_label, color='red', label='average_result_value')
ax1.plot(x_axis, y_res_cl, color='blue', label='res_cl_hist_0min')
ax1.set_ylabel('chlorine')
ax1.set_title("data visualization")

ax2 = ax1.twinx()  # this is the important function
ax2.plot(x_axis, y_zone_cl, 'r', color='green', label='zone_flow_hist_0min')
ax2.set_ylabel('folw')
ax2.set_xlabel('time')

plt.grid(True)
 
plt.legend(bbox_to_anchor=(1.0, 1), loc=1, borderaxespad=0.)

plt.show()

from pandas import DataFrame
df_drop_plot = DataFrame()
df_drop_plot['measurement_result_date'] = df_pump_open['measurement_result_date']
df_drop_plot['average_result_value'] = df_pump_open['average_result_value']
df_drop_plot.plot(x='measurement_result_date', y='average_result_value')
plt.show()

from pandas import DataFrame

df_plot_1 = DataFrame()
df_plot_1= df_pump_open[['average_result_value','res_cl_hist_0min','measurement_result_date','zone_flow_hist_0min']]
df_plot_1.index = df_plot_1['measurement_result_date']
df_plot = df_plot_1.drop(['measurement_result_date'],axis=1)

df_plot.sort_values(['average_result_value']).head(5)

df_plot['average_result_value'].describe()

%matplotlib
import matplotlib.pyplot as plt

df_plot.plot()

df_plot.plot(secondary_y=['zone_flow_hist_0min'])

a = df_plot.sort_values(['average_result_value'])

a[1000:1030]
