==> 03_2 _days _data.ipynb <==
import pandas as pd
from pandas import DataFrame

# input file
output_file_directory = './sample data/output_feature_label.csv'

# output file
output_2_days_directory = './sample data/output_2_days_water_inlet.csv'

df_pump_open = pd.read_csv(output_file_directory, parse_dates=['measurement_result_date'])
df_pump_open.shape

from pandas import DataFrame

df_plot_1 = DataFrame()
df_plot_1= df_pump_open[['average_result_value','res_cl_hist_0min','measurement_result_date','zone_flow_hist_0min']]
df_plot_1.index = df_plot_1['measurement_result_date']
#df_plot = df_plot_1.drop(['measurement_result_date'],axis=1)

df_plot = df_plot_1.sort_index()

df_diff = df_plot.diff()
df_diff_1 = df_diff.drop(['measurement_result_date'],axis=1)
df_diff_1.head(5)

%matplotlib
import matplotlib.pyplot as plt

df_diff_2 = df_diff_1.drop(['zone_flow_hist_0min'],axis=1)
df_diff_2.plot()

df_diff.index = df_diff.index.strftime('%Y-%m-%d %R')

a = df_diff[1020:1081]

df_diff_tem_1 = df_diff.drop(['res_cl_hist_0min', 'zone_flow_hist_0min'],axis=1)
df_diff_tem_1['average_result_value'].describe()

a_1 = a.drop(['zone_flow_hist_0min','res_cl_hist_0min'],axis=1)
# a_1.plot()

a_1

# input: all data
output_all_data = './sample data/output_all_data.csv'

# output: 2 days data
output_2_days_directory = './sample data/output_2_days_water_inlet.csv'

# read all data
data_all = pd.read_csv(output_all_data, parse_dates=['measurement_result_date'])
data_all.shape

data_all = data_all.sort_values(['measurement_result_date'])
data_all.index = data_all['measurement_result_date']

data_2_days = DataFrame()
data_2_days = data_all[['average_result_value','res_cl_hist_0min','measurement_result_date']]
data_2_days.index = data_2_days.index.strftime('%Y-%m-%d %R')

### period 1: 2017-01-02 22:15 ~ 2017-01-03 06:45
* extend close time to 2 hours
* whole open to close period
* 19/09/2017

data_1 = data_2_days['2017-01-02 22:15':'2017-01-03 06:45']
data_1.shape

### period 2: 2017-01-03 22:15 ~ 2017-01-04 06:30

data_2 = data_2_days['2017-01-03 22:15':'2017-01-04 06:30']
data_2.shape

### data: 2个pump open 的时间段 + 关闭后的2个小时

data_2_periods = pd.concat([data_1,data_2])
data_2_periods_df = data_2_periods.drop(['measurement_result_date'],axis=1)
data_2_periods.shape

data_2_periods.to_csv(output_2_days_directory)

data_2_periods_df.info()

### Plot with DataFrame - index is time

data_2_periods_df.plot()

df = data_2_periods

df.head(5)

df.columns = ['value','chlorine','daytime']
df.head(5)

### form proper x-axis for scatter

* need change
* 17/09/2017

df['time'] = df['daytime'].dt.day
df['time'].describe()

df.plot.scatter(x='time',y='value')

data_2.shape
