==> 04_build_dataset.ipynb <==
# data exploration

import pandas as pd

# input file - dataset only when the pump is opening
file_directory = './sample data/output_3_files.csv'

# output file - without useless features (for modelling)
modelling_data_directory = './sample data/modelling_dataset.csv'
modelling_diff_data_directory = './sample data/modelling_diff_dataset.csv'

df = pd.read_csv(file_directory)
df.shape

df.head(5)

## 2 node id in feature dataset

filename = "./sample data/feature.csv"
df_feature = pd.read_csv(filename, parse_dates=['timestamp'])
df_feature['node_id'].describe()

features_32138 = df_feature.loc[df_feature['node_id']==31238]
features_41908 = df_feature.loc[df_feature['node_id']==41908]

### - 判断二个node的数据是否有不同

features_32138['zone_flow_hist_0min'].describe()

features_32138['res_cl_hist_0min'].describe()

features_41908['travel_time_hrs_nom'].describe()

features_41908['travel_time_hrs_nom'].describe()

# data cleaning

df.describe()

df.shape

* average_result_value: label
* Average Result Value: pump speed - sum 3 nodes
* zone_flow_hist_0min: flow speed (record in every 15mins - total: 2 days)
* res_cl_hist_0min: upstream result history of chlorine (record in every 15mins - total: 2 days)
* res_temperature

feature_droped = ['zone_id', 'node_id', 'timestamp', 'flow_nom', 'travel_time_hrs_nom', 'Measurement_Result_Date_Minute_15', 'facility_code', 'asset_code', 'statistic_type', 'measurement_result_date']
df_tem = df.drop(feature_droped ,axis=1)
df_tem.shape

df_tem.head(5)

## output training data
* without timestamp
* 18/09/2017

df_tem.to_csv(modelling_data_directory, index=None)

df_tem.describe()

df_tem.info()

df_tem.head(6)

# Output diff training data
* change label to diff value
* 19/09/2017

df_label_diff = df_tem['average_result_value'].diff()
df_label_diff[2766:]

df_label_diff.describe()

df_tem_1 = df_tem.drop(['average_result_value'], axis=1)
df_diff = pd.concat([df_tem_1, df_label_diff], axis=1)
df_diff.shape

df_diff.head(5)

df_diff.drop([0], inplace=True)

df_diff.head(5)

df_diff.to_csv(modelling_diff_data_directory, index=None)
