==> 06_build_dataset_with_4_files.ipynb <==
# data exploration

import pandas as pd

# input file - dataset only when the pump is opening
file_directory = './sample data/output_all_4_data.csv'

# output file - without useless features (for modelling)
modelling_data_directory = './sample data/modelling_dataset_levels.csv'
modelling_diff_data_directory = './sample data/modelling_diff_dataset_levels.csv'
modelling_only_open_diff_data_directory = './sample data/modelling_only_open_diff_dataset_levels.csv'

df = pd.read_csv(file_directory, parse_dates=['measurement_result_date'])
df.shape

df = df.rename(columns={'Average Result Value': 'pump_average_speed_sum'})

df.head(5)

# data cleaning

df.describe()

* average_result_value: label
* Average Result Value: pump speed - sum 3 nodes
* zone_flow_hist_0min: flow speed (record in every 15mins - total: 2 days)
* res_cl_hist_0min: upstream result history of chlorine (record in every 15mins - total: 2 days)
* res_temperature
* water_level

feature_droped = ['zone_id', 'node_id', 'timestamp', 'flow_nom', 'travel_time_hrs_nom', 'Measurement_Result_Date_Minute_15', 'facility_code', 'asset_code', 'statistic_type']
df_tem = df.drop(feature_droped ,axis=1)
df_tem.shape

# df_tem - droped useless feature data
df_tem.head(5)

# Data pre-processing
* change label to diff value
* 19/09/2017
 - get diff in time
 - 25/09/2018

# get diff in label
df_label_diff = df_tem['average_result_value'].diff()
df_label_diff[8635:]

# label describe
df_label_diff.describe()

# form new diff dataset with all data
df_tem_1 = df_tem.drop(['average_result_value'], axis=1)
df_diff = pd.concat([df_tem_1, df_label_diff], axis=1)
df_diff.shape

df_diff.head(5)

# form new diff dataset with all data
# - without invalid first line

df_diff.drop([0], inplace=True)
df_diff.head(5)

# Output only open dataset

# only open dataset
df_tem_only_open = df_diff.loc[df_diff['pump_average_speed_sum']>0]
df_tem_only_open = df_tem_only_open.sort_values(['measurement_result_date'])

# Output only open dataset
df_tem_only_open.drop(['measurement_result_date'], axis=1).to_csv(modelling_only_open_diff_data_directory, index=None)

# Output open data with 2 hours close

df_tem_only_open.head(4)

df_tem_only_open.shape

# get diff in time
df_time_diff = df_tem_only_open['measurement_result_date'].diff()
df_tem_only_open['time_diff'] = df_time_diff
df_tem_only_open.head(5)

df_tem_only_open['time_diff'].describe()

df_diff = df_diff.sort_values(['measurement_result_date'])

import datetime

count = 0
df_tem_series_result = df_tem_only_open
for i in range(0, len(df_tem_only_open)):
    if df_tem_only_open.iloc[i]['time_diff'] > datetime.timedelta(minutes=15):
#         print df_tem_only_open.iloc[i-1]['measurement_result_date'], "diff:", df_tem_only_open.iloc[i]['time_diff']
        num = df_diff[df_diff['measurement_result_date'] == df_tem_only_open.iloc[i-1]['measurement_result_date']].index.tolist()
#         print num[0]
        df_tem_series = df_diff.iloc[num[0]+1:num[0]+9]
        df_tem_series_result = pd.concat([df_tem_series_result, df_tem_series])

df_tem_series_result.shape

# only open data
df_tem_only_open['measurement_result_date'].describe()

# add 2 hours close data
df_tem_series_result['measurement_result_date'].describe()

# 构建完整，可用于检查的数据 （包含时间戳及时间差）
df_tem_series_result.head(5)

df_tem_series_result.drop_duplicates(['measurement_result_date']).shape

feature_drop = ['measurement_result_date', 'time_diff']
df_series_result = df_tem_series_result.drop_duplicates(['measurement_result_date'])
df_series_result = df_series_result.sort_values(['measurement_result_date'])
df_series_result = df_series_result.drop(feature_drop, axis=1)

df_series_result.shape

# Output add 2h close dataset
df_series_result.to_csv(modelling_diff_data_directory, index=None)

## 总数验证
- 25/09/2018

df_tem_only_open[df_tem_only_open['time_diff'] > datetime.timedelta(minutes=15)].describe()

# 所有 间隔 >15min 的数据 - 需添加 pump 关闭数据的地方
df_tem_only_open[df_tem_only_open['time_diff'] > datetime.timedelta(minutes=15)].shape

# 间隔 >= 2h - 可以完整加入8个数据的地方
df_tem_only_open[df_tem_only_open['time_diff'] >= datetime.timedelta(hours=2)].shape

# 15min < 间隔 < 2h - 加入数据少于8个的地方
df_tem_only_open[(df_tem_only_open['time_diff'] > datetime.timedelta(minutes=15))&(df_tem_only_open['time_diff'] < datetime.timedelta(hours=2))].describe()
